Add customizable prop for spacing Panel items
Both the Palette and the upcoming Toolbar components use the Panel
component to lay out their respective list items but we want each of
the two to have a different gap size (e.g. the Palette's ColorSwatches
packed closer together). This change makes that spacing configurable.

Related to this is the change in the styles of the Panel's list items;
The items are made bigger in width and height to make the clickable
area for interactive items larger. As a result, the ColorSwatch
component is now bigger. It makes it easier to click so this is fine,
just something to note here.