Create Workspace component
The `canvasRef` object is needed by both the Canvas and ClearButton
components for different purposes, so we need to create it in a parent
component and pass it to those two as props. Moved the contents of the
App component into a separate Workspace component so that it makes more
sense than having it created in the top-level App component.